+++ lua/obsidian/config.lua
@@ -68,7 +68,7 @@ end
 
 ---@class obsidian.config.DailyNotesOpts
 ---@field folder string|?
----@field format_date function|?
+---@field date_format string|?
 config.DailyNotesOpts = {}
 
 ---Get defaults.
@@ -76,7 +76,7 @@ config.DailyNotesOpts = {}
 config.DailyNotesOpts.default = function()
   return {
     folder = nil,
-    format_date = nil,
+    date_format = nil,
   }
 end